Move user-course functions to UserController
